requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.49
asyncpg==0.27.0
alembic==1.11.1
pydantic==1.10.8
python-jose==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator==2.2.0
python-multipart==0.0.6
python-socketio>=5.8.0
aiohttp>=3.8.4
websockets>=11.0.2